models and db modules refactored into modules
